refactor: moved client side logic
